Most of the tedious work done in moving things to new Groups.
Some things to figure out/lingering thoughts:
1. The `Group.ts` property makes a kind of copy of the Reader's timestep
   object but with pre-sliced attributes. This feels like it could allow
   some weird stalenss to creep in, and compared to the topology system
   it's very clunky. Need to figure out what to do.
2. Convenience getters like Residue.<atomname> and Segment.<resname>
   amount to __getattr__ modifications for those Components. However,
   these are dependent on TopologyAttrs that are not always present,
   so should be implemented as transplants. But just putting in a
   `__getattr__` method into a `TopologyAttr` overrides its own special
   `__getattr__`. Could just have it exist under a different name,
   but would need special machinery to graft it on under a different
   name. *sigh*
3. Bonds. I leave this to @richardjgowers.
4. Angle conveniences for proteins. Leave in? Or should these be
   deprecated?
