+++ package/MDAnalysis/core/AtomGroup.py
@@ -581,125 +581,6 @@ class AtomGroup(object):
         self.universe.dihedrals += tgroup
         self._clear_caches('dihedrals')
 
-    def wrap(self, compound="atoms", center="com", box=None):
-        """Shift the contents of this AtomGroup back into the unit cell.
-
-        This is a more powerful version of :meth:`pack_into_box`, allowing
-        groups of atoms to be kept together through the process.
-
-        :Keywords:
-           *compound*
-               The group which will be kept together through the shifting
-               process. [``atoms``]
-
-               Possible options:
-
-                   - ``atoms``
-                   - ``group`` - This AtomGroup
-                   - ``residues``
-                   - ``segments``
-                   - ``fragments``
-
-           *center*
-               How to define the center of a given group of atoms [``com``]
-           *box*
-               Unit cell information.  If not provided, the values from
-               Timestep will be used.
-
-        When specifying a *compound*, the translation is calculated based on
-        each compound. The same translation is applied to all atoms
-        within this compound, meaning it will not be broken by the shift.
-        This might however mean that all atoms from the compound are not
-        inside the unit cell, but rather the center of the compound is.
-        Compounds available for use are *atoms*, *residues*,
-        *segments* and *fragments*
-
-        *center* allows the definition of the center of each group to be
-        specified.  This can be either 'com' for center of mass, or 'cog'
-        for center of geometry.
-
-        *box* allows a unit cell to be given for the transformation.  If not
-        specified, an the dimensions information from the current Timestep
-        will be used.
-
-        .. Note::
-           wrap with all default keywords is identical to :meth:`pack_into_box`
-
-        .. versionadded:: 0.9.2
-        """
-        if compound.lower() == "atoms":
-            return self.pack_into_box(box=box)
-
-        if compound.lower() == 'group':
-            objects = [self.atoms]
-        elif compound.lower() == 'residues':
-            objects = self.residues
-        elif compound.lower() == 'segments':
-            objects = self.segments
-        elif compound.lower() == 'fragments':
-            objects = self.fragments
-        else:
-            raise ValueError("Unrecognised compound definition: {0}"
-                             "Please use one of 'group' 'residues' 'segments'"
-                             "or 'fragments'".format(compound))
-
-        if center.lower() in ('com', 'centerofmass'):
-            centers = np.vstack([o.center_of_mass() for o in objects])
-        elif center.lower() in ('cog', 'centroid', 'centerofgeometry'):
-            centers = np.vstack([o.center_of_geometry() for o in objects])
-        else:
-            raise ValueError("Unrecognised center definition: {0}"
-                             "Please use one of 'com' or 'cog'".format(center))
-        centers = centers.astype(np.float32)
-
-        if box is None:
-            box = self.dimensions
-
-        # calculate shift per object center
-        dests = distances.apply_PBC(centers, box=box)
-        shifts = dests - centers
-
-        for o, s in itertools.izip(objects, shifts):
-            # Save some needless shifts
-            if not all(s == 0.0):
-                o.translate(s)
-
-    def split(self, level):
-        """Split atomgroup into a list of atomgroups by *level*.
-
-        *level* can be "atom", "residue", "segment".
-        .. versionadded:: 0.9.0
-        """
-        # CHECK: What happens to duplicate atoms (with advanced slicing)?
-
-        accessors = {'segment': 'segid', 'segid': 'segid',
-                     'residue': 'resid', 'resid': 'resid',
-                     }
-
-        if level == "atom":
-            return [AtomGroup([a]) for a in self]
-
-        # more complicated groupings
-        try:
-            # use own list comprehension to avoid sorting/compression by eg self.resids
-            ids = np.array([getattr(atom, accessors[level]) for atom in self])
-        except KeyError:
-            raise ValueError("level = '{0}' not supported, must be one of {1}".format(
-                    level, accessors.keys()))
-
-        # now sort the resids so that it doesn't matter if the AG contains
-        # atoms in random order (i.e. non-sequential resids); groupby needs
-        # presorted keys!
-        idx = np.argsort(ids)
-        sorted_ids = ids[idx]
-        # group (index, key) and then pull out the index for each group to form AtomGroups
-        # by indexing self (using advanced slicing eg g[[1,2,3]]
-        groups = [
-            self[[idx_k[0] for idx_k in groupings]]  # one AtomGroup for each residue or segment
-            for k, groupings in itertools.groupby(itertools.izip(idx, sorted_ids), lambda v: v[1])
-            ]
-        return groups
-
     def write(self, filename=None, format="PDB",
               filenamefmt="%(trjname)s_%(frame)d", **kwargs):
         """Write AtomGroup to a file.
@@ -814,22 +695,6 @@ class AtomGroup(object):
 
         writer = SelectionWriter(filename, **kwargs)
         writer.write(self.atoms)  # wants a atomgroup
-
-    @property
-    def dimensions(self):
-        """Dimensions of the Universe to which the group belongs, at the current time step."""
-        if self.universe is not None:
-            return self.universe.dimensions
-        else:
-            raise AttributeError("This AtomGroup does not belong to a Universe with a dimension.")
-
-    @dimensions.setter
-    def dimensions(self, box):
-        """Pass on to Universe setter
-
-        .. versionadded:: 0.9.0
-        """
-        self.universe.dimensions = box
 
     @property
     def ts(self):
